// packages/core/src/generated/MediaAsset.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MediaId } from "./MediaId";
import type { MediaKind } from "./MediaKind";
import type { Rate } from "./Rate";
import type { Time } from "./Time";

export type MediaAsset = { id: MediaId, originalName: string, mime: string, kind: MediaKind, sizeBytes: bigint, duration?: Time | null, width?: number | null, height?: number | null, fps?: Rate | null, sampleRate?: number | null, channels?: number | null, };