// packages/core/src/generated/Project.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Marker } from "./Marker";
import type { MasterSettings } from "./MasterSettings";
import type { MediaAsset } from "./MediaAsset";
import type { ProjectMeta } from "./ProjectMeta";
import type { ProjectSettings } from "./ProjectSettings";
import type { Timeline } from "./Timeline";
import type { JsonValue } from "./serde_json/JsonValue";

export type Project = { schemaVersion: number, meta: ProjectMeta, settings: ProjectSettings, library: Array<MediaAsset>, timeline: Timeline, markers: Array<Marker>, master: MasterSettings, } & ({ [key in string]: number | string | boolean | Array<JsonValue> | { [key in string]: JsonValue } | null });